Dev/build oci image for ubuntu 22 04 (#28)
* Add base Dockerfile for Ubuntu-22.04

* Add the Dockerfile to spin up Nginx 1.24.0 and Redis 7.2.1.

This commit introduces a new Dockerfile to the project. This image was built from the base image of ubuntu-22-04.

* Add the Dockerfile to spin up Python 3.12.0

This commit introduces a new Dockerfile to the project. This image was built from the base image of ubuntu-22-04.

* Add the Dockerfile to spin up elasticsearch 8.10.3.

This commit introduces a new Dockerfile to the project. This image was built from the base image of ubuntu-22-04.

* Improve Dockerfile content of nginx-1.24.0 for ubuntu-22-04.dockerfile

* Improve DockerFile of redis-7.2.1 for ubuntu-22-04.dockerfile and add context directory file to nginx-1.24.0 , python-3.12.0 , redis-7.2.1

* Add context directory to elasticsearch-8.10.3

* ADD Context Directory to MongoDB -7.0.2 for Ubuntu-22-04

* Add Dockerfile to spinup Mongodb_7.0.2 for Ubuntu-22-04

* Changes made to improvise Dockerfile of Redis_7.2.1 for Ubuntu_22_04

* Modify docker file to built elasticsearch OCI image for ubuntu-22-04

This commit introduce a new docker file to built elasitcsearch
OCI image for ubuntu-22-04

* Modify docker file to built python 3.12.0 OCI image for ubuntu-22-04

This commit introduces a new Dockerfile to the project.
This image was built from the base image of ubuntu-22-04.

* Bug fix

* Removed unwanted commands
* Fix error in installation command
* Base image name corrected

* Changes made to improvise Dockerfile of Mongodb_7.0.2 for Ubuntu-22-04

---------

Co-authored-by: kumur <kumar.r@contus.in>
Co-authored-by: vinoth kumar <vinothkumar1742@gmail.com>
Co-authored-by: Sujan Kumar <sujankumar4593@gmail.com>